fix hooks locations
