+++ packages/features/data-table/hooks/index.ts
@@ -1,3 +1,6 @@
 export * from "./useDebouncedWidth";
 export * from "./useFetchMoreOnBottomReached";
 export * from "./useColumnSizingVars";
+export * from "./useFilterValue";
+export * from "./useColumnFilters";
+export * from "./useDataTable";
