+++ packages/features/data-table/index.ts
@@ -3,4 +3,4 @@ export * from "./lib/types";
 export * from "./lib/utils";
 export * from "./lib/context";
 export * from "./lib/resizing";
-export * from "./lib/hooks";
+export * from "./hooks";
